ESCLAVO/requirements.txt
cx-Oracle==8.3.0
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
